Refactor layer components into their own files
